Moved import script for tts and vocoder
